fix: shad/cn i added new bg variable to control theme
