+++ app/globals.css
@@ -6,6 +6,7 @@
 :root {
 	/* height: 100svh;
 	width: 100svw; */
+	--myCard: #f0f0f074;
 	--radius: 0.625rem;
 	--background: oklch(1 0 0);
 	--foreground: oklch(0.129 0.042 264.695);
@@ -74,6 +75,7 @@
   --color-popover: var(--popover);
   --color-card-foreground: var(--card-foreground);
   --color-card: var(--card);
+  --color-myCard: var(--myCard);
   --radius-sm: calc(var(--radius) - 4px);
   --radius-md: calc(var(--radius) - 2px);
   --radius-lg: var(--radius);
@@ -84,6 +86,8 @@
   :root {
     --background: #121214;
 	--foreground: #f5f5f5;
+	/* --card: #252427; */
+	--myCard: #252427;
   }
 }
 
@@ -129,6 +133,7 @@ body {
 	--sidebar-accent-foreground: oklch(0.984 0.003 247.858);
 	--sidebar-border: oklch(1 0 0 / 10%);
 	--sidebar-ring: oklch(0.551 0.027 264.364);
+	--myCard: #252427;
 }
 
 @layer base {
