+++ presentation/components/LanguageSelector.tsx
@@ -44,24 +44,24 @@ const LanguageSelector: React.FC<LanguageSelectorProps> = ({
 
 	return (
 		<div className="px-3 md:px-4 w-full">
-			<div className="flex h-16 md:h-18 items-center justify-start border-b border-gray-700 bg-[#252427] px-3 md:px-6 py-2 rounded-lg">
+			<div className="flex h-16 md:h-18 items-center justify-start border-b border-gray-700 bg-myCard px-3 md:px-6 py-2 rounded-lg">
 				<div className="flex w-full h-full items-center gap-2 md:gap-4">
 					<LanguageSelect
 						value={fromLanguage}
 						setValue={setFromLanguage}
 						disabledValue={toLanguage}
-						className="flex-1"
+						className="flex-1 text-foreground"
 					/>
 					<Button
 						variant="ghost"
 						size="icon"
-						className="h-full min-w-[40px] md:min-w-[80px] hover:bg-gray-700"
+						className="h-full text-foreground min-w-[40px] md:min-w-[80px] hover:bg-gray-700"
 						onClick={handleSwapLanguages}
 					>
 						<ArrowLeftRight
 							height={32}
 							width={16}
-							className="text-gray-400 stroke-2 size-4 md:size-6"
+							className="text-gray-400 text-foreground stroke-2 size-4 md:size-6"
 						/>
 					</Button>
 					<div className="w-full h-full flex flex-1 items-center">
@@ -69,10 +69,10 @@ const LanguageSelector: React.FC<LanguageSelectorProps> = ({
 							value={toLanguage}
 							setValue={setToLanguage}
 							disabledValue={fromLanguage}
-							className="flex-1"
+							className="flex-1 text-foreground"
 						/>
-						<div className="text-xs p-0 md:text-base md:mx-2 h-full w-[1px] bg-gray-700"></div>
-						<ToneSelector value={tone} onToneChange={onToneChange} />
+						<div className="text-xs p-0 md:text-base md:mx-2 h-full w-[1px] bg-foreground"></div>
+						<ToneSelector value={tone} onToneChange={onToneChange} className="text-foreground" />
 					</div>
 				</div>
 			</div>
