+++ shared/constants/languages.ts
@@ -1,22 +1,2 @@
-export const languages = {
-	en: {
-		label: "English",
-		code: "en",
-	},
-	ru: {
-		label: "Russian",
-		code: "ru",
-	},
-	es: {
-		label: "Spanish",
-		code: "es",
-	},
-	fr: {
-		label: "French",
-		code: "fr",
-	},
-	de: {
-		label: "German",
-		code: "de",
-	}
-} as const;
+// Re-export domain constants for presentation layer compatibility
+export { languages } from "../../domain/constants/languages";
